Use domain relative urls
Less verbose html pages from it, and ability to test on a different domain. In addition the `relativeURLs` site config will allow it to work just from the file-system.

The only remaining absolute urls are either meant to be (as is for RSS feeds), or the html RSSLink. The latter there is no obvious way to make relative.