Replace social icons with  Font Awesome icons
There is no need to generate svg path when adding svg icons. So that we canI easily add more icons to the theme, and the style is uniform. Font Awesome gives you scalable vector icons that can instantly be customized — size, color, drop shadow, and anything that can be done with the power of CSS. http://fontawesome.io/