Display the post author in the post meta informations.
In config.toml, the key post_meta defines a range of meta informations that can
be displayed under the post title. This commit allows to display the author name
in these meta informations, by adding the constant "author" in the post_meta
range, like so:

post_meta = ["author", "date", "categories", "translations"]

This meta information will be displayed only if an author name is defined.
Moreover, if the Param authorbox is set to true in config.toml, the name is
displayed as a link to the authorbox on the same page.