Fix a bug when displaying the author name as post meta.
When displaying the author post meta (author.html), don't create a link if the
partial is called from the summary (summary.html). The link created by
author.html points to the authorbox, which exists only when displaying a
post page, so we don't need to create a link to it in the summary.